+++ app.go
@@ -1205,11 +1205,12 @@ func (e *Error) Error() string {
 
 // NewError creates a new Error instance with an optional message
 func NewError(code int, message ...string) *Error {
-	err := &Error{Code: code}
+	err := &Error{
+		Code:    code,
+		Message: utils.StatusMessage(code),
+	}
 	if len(message) > 0 {
 		err.Message = message[0]
-	} else {
-		err.Message = utils.StatusMessage(code)
 	}
 	return err
 }
